+++ test/models/px/test_fuxi.py
@@ -22,33 +22,52 @@ import pytest
 import torch
 
 from earth2studio.data import Random, fetch_data
-from earth2studio.models.px import DLWP
+from earth2studio.models.auto import Package
+from earth2studio.models.px import FuXi
 from earth2studio.utils import handshake_dim
 
 
-class PhooDLWPModel(torch.nn.Module):
-    """Dummy DLWP model, adds time-step"""
+class PhooFuXiModel(torch.nn.Module):
+    """Dummy FuXi model, adds time-step"""
 
-    def __init__(self, delta_t: int = 6):
+    def __init__(self, model_type: str = "short"):
         super().__init__()
-        self.delta_t = delta_t
+        # Model cascade testing
+        if model_type == "short":
+            self.delta_t = 1
+        elif model_type == "medium":
+            self.delta_t = 2
+        else:
+            self.delta_t = 3
 
-    def forward(self, x):
-        # 8:15 because field 7 is cosine zenith of first timestep
-        x[:, :7] = x[:, 8:15] + self.delta_t
-        x[:, 7:14] = x[:, 8:15] + 2 * self.delta_t
-        return x[:, :14].contiguous()
+    def forward(self, x, y):
+        # Remove first time-step
+        assert y.shape[1] == 12
+        # Add  0*y[0,0] so input y remains in ONNX graph
+        output = x + self.delta_t + 0 * y[0, 0]
+        return output
 
 
-@pytest.fixture()
-def dlwp_phoo_cs_transform():
-    er_num = 721 * 1440
-    cs_num = 6 * 64 * 64
-    values = np.ones(cs_num)
-    indices = np.stack([np.arange(cs_num), np.arange(cs_num)], axis=0)
-    return torch.sparse_coo_tensor(
-        indices, values, size=(cs_num, er_num), dtype=torch.float
-    )
+@pytest.fixture(scope="session")
+def fuxi_test_package(tmp_path_factory):
+    """Creates a bunch of spoof ONNX models to unit test with"""
+    tmp_path = tmp_path_factory.mktemp("data")
+
+    for model in ["short", "medium", "long"]:
+        onnx_path = tmp_path / f"{model}.onnx"
+        torch.onnx.export(
+            PhooFuXiModel(model_type=model),
+            args=(torch.rand(1, 2, 70, 721, 1440), torch.rand(1, 12)),
+            f=str(onnx_path),
+            export_params=True,
+            opset_version=10,
+            input_names=["input", "temb"],
+            output_names=["output"],
+        )
+        # Empty weight file
+        open(tmp_path / f"{model}", "a").close()
+
+    return Package(str(tmp_path))
 
 
 @pytest.mark.parametrize(
@@ -64,25 +83,10 @@ def dlwp_phoo_cs_transform():
     ],
 )
 @pytest.mark.parametrize("device", ["cpu", "cuda:0"])
-def test_dlwp_call(time, dlwp_phoo_cs_transform, device):
-    model = PhooDLWPModel()
-    landsea_mask = torch.ones(6, 64, 64)
-    orography = torch.ones(6, 64, 64)
-    latgrid = torch.ones(6, 64, 64)
-    longrid = torch.ones(6, 64, 64)
-    center = torch.zeros(1, 7, 1, 1)
-    scale = torch.ones(1, 7, 1, 1)
-    p = DLWP(
-        model,
-        landsea_mask=landsea_mask,
-        orography=orography,
-        latgrid=latgrid,
-        longrid=longrid,
-        cubed_sphere_transform=dlwp_phoo_cs_transform,
-        cubed_sphere_inverse=dlwp_phoo_cs_transform.T,
-        center=center,
-        scale=scale,
-    ).to(device)
+def test_fuxi_call(time, fuxi_test_package, device):
+
+    # Use dummy package
+    p = FuXi.load_model(fuxi_test_package).to(device)
 
     dc = p.input_coords.copy()
     del dc["batch"]
@@ -108,8 +112,8 @@ def test_dlwp_call(time, dlwp_phoo_cs_transform, device):
     assert (out_coords["variable"] == p.output_coords["variable"]).all()
     assert (out_coords["time"] == time).all()
     assert torch.allclose(
-        out, p.to_equirectangular(p.to_cubedsphere(x[:, 1:] + 6))
-    )  # Need to cs transform here to get right values
+        out, (x[:, 1:] + 1)
+    )  # Phoo model should add by delta t each call
     handshake_dim(out_coords, "lon", 4)
     handshake_dim(out_coords, "lat", 3)
     handshake_dim(out_coords, "variable", 2)
@@ -119,30 +123,13 @@ def test_dlwp_call(time, dlwp_phoo_cs_transform, device):
 
 @pytest.mark.parametrize(
     "ensemble",
-    [1, 2],
+    [2],
 )
 @pytest.mark.parametrize("device", ["cpu", "cuda"])
-def test_dlwp_iter(ensemble, dlwp_phoo_cs_transform, device):
+def test_fuxi_iter(ensemble, fuxi_test_package, device):
     time = np.array([np.datetime64("1993-04-05T00:00")])
-    # Use dummy model
-    model = PhooDLWPModel()
-    landsea_mask = torch.ones(6, 64, 64)
-    orography = torch.ones(6, 64, 64)
-    latgrid = torch.ones(6, 64, 64)
-    longrid = torch.ones(6, 64, 64)
-    center = torch.zeros(1, 7, 1, 1)
-    scale = torch.ones(1, 7, 1, 1)
-    p = DLWP(
-        model,
-        landsea_mask=landsea_mask,
-        orography=orography,
-        latgrid=latgrid,
-        longrid=longrid,
-        cubed_sphere_transform=dlwp_phoo_cs_transform,
-        cubed_sphere_inverse=dlwp_phoo_cs_transform.T,
-        center=center,
-        scale=scale,
-    ).to(device)
+    # Use dummy package
+    p = FuXi.load_model(fuxi_test_package).to(device)
 
     dc = p.input_coords.copy()
     del dc["batch"]
@@ -169,18 +156,26 @@ def test_dlwp_iter(ensemble, dlwp_phoo_cs_transform, device):
 
     # Get generator
     out, out_coords = next(p_iter)  # Skip first which should return the input
-    assert torch.allclose(out, x[:, :, 1:])
+    assert torch.allclose(out, x[:, 1:])
 
+    step_index = 0
     for i, (out, out_coords) in enumerate(p_iter):
+        # Test the model cascade
+        if i < 20:
+            step_index += 1
+        elif i < 40:
+            step_index += 2
+        else:
+            step_index += 3
+
         assert len(out.shape) == 6
         assert out.shape[0] == ensemble
         assert (out_coords["variable"] == p.output_coords["variable"]).all()
         assert (out_coords["time"] == time).all()
         assert out_coords["lead_time"][0] == np.timedelta64(6 * (i + 1), "h")
         assert torch.allclose(
-            out, p.to_equirectangular(p.to_cubedsphere(x[:, :, 1:] + 6 * (i + 1)))
-        )  # Need to cs transform here to get right values
-
+            out, (x[:, 1:] + step_index)
+        )  # Phoo model should add by delta t each call
         handshake_dim(out_coords, "lon", 5)
         handshake_dim(out_coords, "lat", 4)
         handshake_dim(out_coords, "variable", 3)
@@ -188,41 +183,24 @@ def test_dlwp_iter(ensemble, dlwp_phoo_cs_transform, device):
         handshake_dim(out_coords, "time", 1)
         handshake_dim(out_coords, "ensemble", 0)
 
-        if i > 5:
+        if i > 41:  # Long test because of model cascade
             break
 
 
 @pytest.mark.parametrize(
     "dc",
     [
-        OrderedDict({"lat": np.random.randn(720)}),
-        OrderedDict({"lat": np.random.randn(720), "phoo": np.random.randn(1440)}),
-        OrderedDict({"lat": np.random.randn(720), "lon": np.random.randn(1)}),
+        OrderedDict({"lat": np.random.randn(721)}),
+        OrderedDict({"lat": np.random.randn(721), "phoo": np.random.randn(1440)}),
+        OrderedDict({"lat": np.random.randn(721), "lon": np.random.randn(1)}),
     ],
 )
 @pytest.mark.parametrize("device", ["cuda:0"])
-def test_dlwp_exceptions(dc, dlwp_phoo_cs_transform, device):
+def test_fuxi_exceptions(dc, fuxi_test_package, device):
     # Test invalid coordinates error
     time = np.array([np.datetime64("1993-04-05T00:00")])
-    # Use dummy model
-    model = PhooDLWPModel()
-    landsea_mask = torch.ones(6, 64, 64)
-    orography = torch.ones(6, 64, 64)
-    latgrid = torch.ones(6, 64, 64)
-    longrid = torch.ones(6, 64, 64)
-    center = torch.zeros(1, 7, 1, 1)
-    scale = torch.ones(1, 7, 1, 1)
-    p = DLWP(
-        model,
-        landsea_mask=landsea_mask,
-        orography=orography,
-        latgrid=latgrid,
-        longrid=longrid,
-        cubed_sphere_transform=dlwp_phoo_cs_transform,
-        cubed_sphere_inverse=dlwp_phoo_cs_transform.T,
-        center=center,
-        scale=scale,
-    ).to(device)
+    # Use dummy package
+    p = FuXi.load_model(fuxi_test_package).to(device)
 
     # Initialize Data Source
     r = Random(dc)
@@ -239,12 +217,12 @@ def test_dlwp_exceptions(dc, dlwp_phoo_cs_transform, device):
 @pytest.mark.ci_cache
 @pytest.mark.timeout(360)
 @pytest.mark.parametrize("device", ["cuda:0"])
-def test_dlwp_package(device, model_cache_context):
+def test_fuxi_package(device, model_cache_context):
     time = np.array([np.datetime64("1993-04-05T00:00")])
     with model_cache_context():
         with torch.device(device):
-            package = DLWP.load_default_package()
-            p = DLWP.load_model(package).to(device)
+            package = FuXi.load_default_package()
+            p = FuXi.load_model(package).to(device)
 
     dc = p.input_coords.copy()
     del dc["batch"]
@@ -264,7 +242,9 @@ def test_dlwp_package(device, model_cache_context):
     if not isinstance(time, Iterable):
         time = [time]
 
-    assert out.shape == torch.Size([len(time), 1, 7, 721, 1440])
+    assert out.shape == torch.Size(
+        [len(time), 1, len(p.output_coords["variable"]), 721, 1440]
+    )
     assert (out_coords["variable"] == p.output_coords["variable"]).all()
     assert (out_coords["time"] == time).all()
     handshake_dim(out_coords, "lon", 4)